move inventory related functions into inventory plugin
