+++ imports/plugins/core/catalog/server/no-meteor/utils/updateCatalogProductInventoryStatus.js
@@ -1,8 +1,8 @@
 import Logger from "@reactioncommerce/logger";
 import _ from "lodash";
-import isBackorder from "./isBackorder";
-import isLowQuantity from "./isLowQuantity";
-import isSoldOut from "./isSoldOut";
+import isBackorder from "/imports/plugins/core/inventory/server/no-meteor/utils/isBackorder";
+import isLowQuantity from "/imports/plugins/core/inventory/server/no-meteor/utils/isLowQuantity";
+import isSoldOut from "/imports/plugins/core/inventory/server/no-meteor/utils/isSoldOut";
 
 /**
  *
